adds particle interaction-range table
NOTES:
The table defines the particle-particle interaction range given
the kinds of the pair. Pairs beyond the interaction range
experience a zero force from one another.

Adds only for spheres since it is the kind that I am currently
working in earnest.

adds another entry to `enum kind` for the purposes of knowing
how many enums have been defined; based on stackoverflow there's no
consensus on what's the best approach so I picked the one that
I fancied. (Maybe some cool feature shall be added to C++ that solves
this in a more elegant manner but for now this is "good enough" for
this case.)